feat: Add Time component for selecting movie duration
The commit adds a new component called Time, which allows users to select the duration of the movie they would like to watch. This enhances the user experience by providing a customizable time selection feature. The Time component is implemented in the src/components/time.tsx file and includes a slider for selecting different time intervals. This addition improves the functionality and interactivity of the application.